requirements.txt
asgiref==3.7.1
async-generator==1.10
attrs==23.1.0
autopep8==2.0.2
certifi==2023.5.7
charset-normalizer==3.1.0
Django==4.2.1
django-autoslug==1.9.9
django-debug-toolbar==4.1.0
django-silk==0.5.2
djangorestframework==3.14.0
exceptiongroup==1.1.1
factory-boy==3.2.1
Faker==18.9.0
h11==0.14.0
idna==3.4
inflection==0.5.1
iniconfig==2.0.0
Jinja2==3.1.2
MarkupSafe==2.1.2
outcome==1.2.0
packaging==23.1
pluggy==1.0.0
py==1.11.0
pycodestyle==2.10.0
Pygments==2.15.1
PySocks==1.7.1
pytest==6.2.5
pytest-base-url==2.0.0
pytest-django==4.5.2
pytest-factoryboy==2.5.1
pytest-html==3.2.0
pytest-metadata==2.0.4
pytest-selenium==4.0.0
pytest-variables==2.0.0
python-dateutil==2.8.2
pytz==2023.3
requests==2.31.0
selenium==4.9.1
simplejson==3.19.1
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
sqlparse==0.4.4
tenacity==6.3.1
toml==0.10.2
tomli==2.0.1
trio==0.22.0
trio-websocket==0.10.2
typing_extensions==4.6.2
urllib3==2.0.2
wsproto==1.2.0
